Move inner class to upper level
